requirements.txt
aliyun-python-sdk-core==2.13.4
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-kms==2.6.0
amqp==2.5.0
asn1crypto==0.24.0
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
dingtalk-sdk==1.2.8
Django==2.0
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-guardian==2.0.0
django-import-export==1.2.0
django-redis-cache==2.0.0
django-redis-sessions==0.6.1
django-restframework-apiview==1.3.6
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
et-xmlfile==1.0.1
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
httplib2==0.12.3
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.1
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
oss2==2.6.1
Pillow==6.0.0
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
PyJWT==1.7.1
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.22.0
six==1.12.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
